fix: seperate the validators intoa separate file and added lottie for representing no data
