+++ lib/services/auth_service.dart
@@ -1,10 +1,13 @@
 import 'package:firebase_auth/firebase_auth.dart';
 
+// Service class for handling authentication with Firebase.
 class AuthService {
   final FirebaseAuth _auth = FirebaseAuth.instance;
 
+  // Stream of user authentication state changes.
   Stream<User?> get user => _auth.authStateChanges();
 
+  // Signs in user with email and password.
   Future<User?> signIn(String email, String password) async {
     try {
       final cred = await _auth.signInWithEmailAndPassword(
@@ -17,6 +20,7 @@ class AuthService {
     }
   }
 
+  // Registers a new user with email and password.
   Future<User?> register(String email, String password) async {
     try {
       final cred = await _auth.createUserWithEmailAndPassword(
@@ -29,6 +33,7 @@ class AuthService {
     }
   }
 
+  // Sends password reset email.
   Future<void> sendPasswordReset(String email) async {
     try {
       await _auth.sendPasswordResetEmail(email: email.trim());
@@ -37,6 +42,7 @@ class AuthService {
     }
   }
 
+  // Signs out the current user.
   Future<void> signOut() async {
     await _auth.signOut();
   }
