+++ lib/viewmodels/task_viewmodel.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import '../models/task.dart';
 import '../services/task_repository.dart';
 
+// ViewModel for managing tasks, filters, and state.
 class TaskViewModel extends ChangeNotifier {
   late TaskRepository _repo;
   List<Task> _allTasks = [];
@@ -15,6 +16,7 @@ class TaskViewModel extends ChangeNotifier {
   bool get isBusy => _isBusy;
   bool get isLoading => _isLoading;
 
+  // Filtered tasks by status.
   List<Task> get todoTasks =>
       _filteredTasks.where((t) => t.status == 'todo').toList();
   List<Task> get pendingTasks =>
@@ -22,6 +24,7 @@ class TaskViewModel extends ChangeNotifier {
   List<Task> get completedTasks =>
       _filteredTasks.where((t) => t.status == 'completed').toList();
 
+  // Completion progress percentage.
   double get completionProgress =>
       _allTasks.isEmpty ? 0.0 : completedTasks.length / _allTasks.length;
 
@@ -29,13 +32,14 @@ class TaskViewModel extends ChangeNotifier {
 
   StreamSubscription<List<Task>>? _subscription;
 
+  // Initializes repository and starts listening to tasks stream.
   void init(String userId) {
     _repo = TaskRepository(userId: userId);
     _isLoading = true;
     notifyListeners();
     _subscription?.cancel();
     _subscription = _repo.getTasks().listen(
-      (tasks) {
+          (tasks) {
         _allTasks = tasks;
         _applyFilters();
         _isLoading = false;
@@ -48,6 +52,7 @@ class TaskViewModel extends ChangeNotifier {
     );
   }
 
+  // Refreshes tasks by fetching once.
   Future<void> refresh() async {
     try {
       final tasks = await _repo.fetchTasksOnce();
@@ -57,24 +62,28 @@ class TaskViewModel extends ChangeNotifier {
     } catch (_) {}
   }
 
+  // Sets search query and applies filters.
   void setSearchQuery(String q) {
     _searchQuery = q.toLowerCase();
     _applyFilters();
     notifyListeners();
   }
 
+  // Clears search query.
   void clearSearch() {
     _searchQuery = '';
     _applyFilters();
     notifyListeners();
   }
 
+  // Sets sort method and applies filters.
   void setSortBy(String sort) {
     _sortBy = sort;
     _applyFilters();
     notifyListeners();
   }
 
+  // Applies search and sort filters to tasks.
   void _applyFilters() {
     _filteredTasks = _allTasks.where((t) {
       final matchesQuery =
@@ -100,6 +109,7 @@ class TaskViewModel extends ChangeNotifier {
     }
   }
 
+  // Adds a task with busy state handling.
   Future<void> addTask(Task task) async {
     _setBusy(true);
     try {
@@ -109,6 +119,7 @@ class TaskViewModel extends ChangeNotifier {
     }
   }
 
+  // Updates a task with busy state handling.
   Future<void> updateTask(Task task) async {
     _setBusy(true);
     try {
@@ -118,6 +129,7 @@ class TaskViewModel extends ChangeNotifier {
     }
   }
 
+  // Updates task status with busy state handling.
   Future<void> updateStatus(Task task, String status) async {
     _setBusy(true);
     try {
@@ -127,6 +139,7 @@ class TaskViewModel extends ChangeNotifier {
     }
   }
 
+  // Deletes a task with busy state handling.
   Future<void> deleteTask(String id) async {
     _setBusy(true);
     try {
@@ -136,6 +149,7 @@ class TaskViewModel extends ChangeNotifier {
     }
   }
 
+  // Toggles completion with busy state handling.
   Future<void> toggleCompletion(Task task) async {
     _setBusy(true);
     try {
@@ -145,6 +159,7 @@ class TaskViewModel extends ChangeNotifier {
     }
   }
 
+  // Sets busy state and notifies listeners.
   void _setBusy(bool v) {
     _isBusy = v;
     notifyListeners();
